appConfig now handles settings as its own module
